+++ src/main/java/com/asankasi/dronefleet/service/MedicationService.java
@@ -5,5 +5,5 @@ import com.asankasi.dronefleet.model.Medication;
 public interface MedicationService {
     Medication findMedication(Long id);
 
-    Double getMinUnitWeight();
+    Integer getMinUnitWeight();
 }
